An unreasonable amount of work for slightly better translation support...
