+++ src/components/App.jsx
@@ -1,101 +1,25 @@
 import '../assets/scss/main.scss'
 
-import React, { Suspense, useEffect, useState, useCallback } from 'react'
+import React, { Suspense } from 'react'
 import { ApolloProvider } from '@apollo/client'
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom'
-import { ThemeProvider } from '@material-ui/styles'
-
-import setTheme from '@assets/mui/theme'
-import UIcons from '@services/Icons'
-import Fetch from '@services/Fetch'
 import client from '@services/apollo'
 
-import Auth from './layout/auth/Auth'
-import Login from './layout/auth/Login'
-import RouteChangeTracker from './RouteChangeTracker'
-import Errors from './Errors'
-import ClearStorage from './ClearStorage'
-import HolidayEffects from './HolidayEffects'
+import ReactRouter from './ReactRouter'
+
+const SetText = () => {
+  const loadingText = document.getElementById('loading-text')
+  if (loadingText) loadingText.innerHTML = 'Loading Translations'
+  return <></>
+}
 
 export default function App() {
-  const [serverSettings, setServerSettings] = useState(null)
-
-  const rootLoading = document.getElementById('loader')
-  const loadingText = document.getElementById('loading-text')
-  if (rootLoading) {
-    if (loadingText) {
-      loadingText.innerHTML = 'Loading Translations'
-    }
-    if (serverSettings) {
-      rootLoading.style.display = 'none'
-    }
-  }
-
-  const getServerSettings = useCallback(async () => {
-    const data = await Fetch.getSettings()
-    const Icons = data.masterfile ? new UIcons(data.config.icons, data.masterfile.questRewardTypes) : null
-    if (Icons) {
-      if (loadingText) {
-        loadingText.innerHTML = 'Loading Icons'
-      }
-      await Icons.fetchIcons(data.config.icons.styles)
-      if (data.config.icons.defaultIcons) {
-        Icons.setSelection(data.config.icons.defaultIcons)
-      }
-    }
-    if (data.ui?.pokestops?.invasions && data.config?.map.fetchLatestInvasions) {
-      const invasionCache = JSON.parse(localStorage.getItem('invasions_cache'))
-      const cacheTime = data.config.map.invasionCacheHrs * 60 * 60 * 1000
-      if (invasionCache && invasionCache.lastFetched + cacheTime > Date.now()) {
-        data.masterfile.invasions = invasionCache
-      } else {
-        if (loadingText) {
-          loadingText.innerHTML = 'Loading Invasions'
-        }
-        data.masterfile.invasions = await Fetch.getInvasions(data.masterfile.invasions)
-      }
-    }
-    setServerSettings({ ...data, Icons })
-  }, [])
-
-  useEffect(() => {
-    getServerSettings()
-  }, [])
+  document.body.classList.add('dark')
 
   return (
-    <ThemeProvider theme={setTheme(serverSettings?.config?.map?.theme)}>
-      <Suspense fallback="">
-        <ApolloProvider client={client}>
-          <Router>
-            {(process.env && process.env.GOOGLE_ANALYTICS_ID) && <RouteChangeTracker />}
-            <Switch>
-              <Route exact path="/404" component={Errors} />
-              <Route exact path="/500" component={Errors} />
-              <Route exact path="/reset" component={ClearStorage} />
-              <Route exact path="/">
-                {serverSettings && <Auth serverSettings={serverSettings} getServerSettings={getServerSettings} />}
-              </Route>
-              <Route exact path="/login">
-                {serverSettings && (
-                  <Login
-                    clickedTwice
-                    serverSettings={serverSettings}
-                    getServerSettings={getServerSettings}
-                  />
-                )}
-              </Route>
-              <Route exact path="/@/:lat/:lon/:zoom?">
-                {serverSettings && <Auth serverSettings={serverSettings} getServerSettings={getServerSettings} />}
-              </Route>
-              <Route exact path="/id/:category/:id/:zoom?">
-                {serverSettings && <Auth serverSettings={serverSettings} getServerSettings={getServerSettings} />}
-              </Route>
-            </Switch>
-          </Router>
-          <canvas id="holiday-canvas" />
-          <HolidayEffects mapSettings={serverSettings?.config?.map ? serverSettings.config.map : {}} />
-        </ApolloProvider>
-      </Suspense>
-    </ThemeProvider>
+    <Suspense fallback={<SetText />}>
+      <ApolloProvider client={client}>
+        <ReactRouter />
+      </ApolloProvider>
+    </Suspense>
   )
 }
