+++ src/components/ConfigSettings.jsx
@@ -15,9 +15,6 @@ export default function ConfigSettings({
 }) {
   Utility.analytics('User', serverSettings.user ? `${serverSettings.user.username} (${serverSettings.user.id})` : 'Not Logged In', 'Permissions', true)
 
-  document.title = serverSettings.config.map.headerTitle
-  document.body.classList.add('dark')
-
   const setUserSettings = useStore(state => state.setUserSettings)
   const setSettings = useStore(state => state.setSettings)
   const setFilters = useStore(state => state.setFilters)
